+++ scale_figure/__init__.py
@@ -1,9 +1,11 @@
 from . import (
     draw_scale_figure,
+    bake_figure,
     )
 
 modules = (
     draw_scale_figure,
+    bake_figure
 )
 
 def register():